wifi: Move assert to correct position
When an HT station sends a (HR-)DSSS PPDU, the bandwidth used by the PPDU (22 MHz)
may exceed the channel width (20 MHz). This assert was intended to check that the
width of the primary channel used to transmit a PPDU does not exceed the width of
the operating channel, hence this check does not apply to (HR-)DSSS PPDUs.