wifi: handle modulation not supported in StartRx
This commit avoids triggering the abort in WifiPhy::GetPhyEntity in case
a device receives a signal using a modulation class that is not
supported. We achieve this by not calling WifiPhy::GetPhyEntity
altogether: checks related to sensitivity and overlap with the primary
20 MHz channel are now performed without relying on PhyEntity, thanks to
a refactoring moving GetTransmissionChannelWidth from PhyEntity to
WifiPpdu and moving PhyEntity::CanReceivePpdu to
WifiPpdu::CanBeReceived.