wifi: Do not check lifetime expiration when dequeuing frames
Frames are dequeued when they have been acknowledged, dropped
or are going to be replaced by other frames. In all such cases,
it is not needed to check for lifetime expiration. This check instead
causes troubles when the lifetime of a frame expires while receiving
the acknowledgment (a BAR may be scheduled unnecessarily).