wifi: Reset current MU PPDU UID when payload reception ends
Improves the fix applied with bedf16d4. Basically, a non-AP STA could
receive a DL MU PPDU including a MU-BAR TF, store the MU PPDU UID when
processing SIG-B, do not respond to the MU-BAR TF (this may happen in
some cases with EMLSR clients) and used the stored MU PPDU UID to schedule
the reception of the OFDMA payload of the TB PPDU sent by other non-AP STAs.